fix missing enterprisedb postgres bundles for some platforms

+++ repack-postgres.sh
@@ -20,9 +20,12 @@ mkdir -p $DIST_DIR $TRG_DIR
 
 [ -e $OSX_DIST ] || wget -O $OSX_DIST "http://get.enterprisedb.com/postgresql/postgresql-$VERSION-osx-binaries.zip"
 [ -e $WINDOWS_DIST ] || wget -O $WINDOWS_DIST "http://get.enterprisedb.com/postgresql/postgresql-$VERSION-windows-x64-binaries.zip"
-[ -e $WINDOWS_X32_DIST ] || wget -O $WINDOWS_X32_DIST "http://get.enterprisedb.com/postgresql/postgresql-$VERSION-windows-binaries.zip"
-[ -e $LINUX_DIST ] || wget -O $LINUX_DIST "http://get.enterprisedb.com/postgresql/postgresql-$VERSION-linux-x64-binaries.tar.gz"
-[ -e $LINUX_X32_DIST ] || wget -O $LINUX_X32_DIST "http://get.enterprisedb.com/postgresql/postgresql-$VERSION-linux-binaries.tar.gz"
+
+if [[ "$VERSION" == 9.* ]] || [[ "$VERSION" == 10.* ]] ; then
+  [ -e $WINDOWS_X32_DIST ] || wget -O $WINDOWS_X32_DIST "http://get.enterprisedb.com/postgresql/postgresql-$VERSION-windows-binaries.zip"
+  [ -e $LINUX_DIST ] || wget -O $LINUX_DIST "http://get.enterprisedb.com/postgresql/postgresql-$VERSION-linux-x64-binaries.tar.gz"
+  [ -e $LINUX_X32_DIST ] || wget -O $LINUX_X32_DIST "http://get.enterprisedb.com/postgresql/postgresql-$VERSION-linux-binaries.tar.gz"
+fi
 
 rm -rf $PKG_DIR/darwin && mkdir -p $PKG_DIR/darwin
 unzip -q -d $PKG_DIR/darwin $OSX_DIST
@@ -54,37 +57,41 @@ tar -cJvf $TRG_DIR/postgres-windows-x86_64.txz \
   bin/postgres.exe \
   bin/*.dll
 
-rm -rf $PKG_DIR/windows_x32 && mkdir -p $PKG_DIR/windows_x32
-unzip -q -d $PKG_DIR/windows_x32 $WINDOWS_X32_DIST
-cd $PKG_DIR/windows_x32/pgsql
-tar -cJvf $TRG_DIR/postgres-windows-x86_32.txz \
-  share \
-  lib/iconv.lib \
-  lib/libxml2.lib \
-  lib/ssleay32.lib \
-  lib/ssleay32MD.lib \
-  lib/*.dll \
-  bin/initdb.exe \
-  bin/pg_ctl.exe \
-  bin/postgres.exe \
-  bin/*.dll
+if [[ "$VERSION" == 9.* ]] || [[ "$VERSION" == 10.* ]] ; then
 
-rm -rf $PKG_DIR/linux && mkdir -p $PKG_DIR/linux
-tar -xzf $LINUX_DIST -C $PKG_DIR/linux
-cd $PKG_DIR/linux/pgsql
-tar -cJvf $TRG_DIR/postgres-linux-x86_64.txz \
-  share/postgresql \
-  lib \
-  bin/initdb \
-  bin/pg_ctl \
-  bin/postgres
+  rm -rf $PKG_DIR/windows_x32 && mkdir -p $PKG_DIR/windows_x32
+  unzip -q -d $PKG_DIR/windows_x32 $WINDOWS_X32_DIST
+  cd $PKG_DIR/windows_x32/pgsql
+  tar -cJvf $TRG_DIR/postgres-windows-x86_32.txz \
+    share \
+    lib/iconv.lib \
+    lib/libxml2.lib \
+    lib/ssleay32.lib \
+    lib/ssleay32MD.lib \
+    lib/*.dll \
+    bin/initdb.exe \
+    bin/pg_ctl.exe \
+    bin/postgres.exe \
+    bin/*.dll
 
-rm -rf $PKG_DIR/linux_x32 && mkdir -p $PKG_DIR/linux_x32
-tar -xzf $LINUX_X32_DIST -C $PKG_DIR/linux_x32
-cd $PKG_DIR/linux_x32/pgsql
-tar -cJvf $TRG_DIR/postgres-linux-x86_32.txz \
-  share/postgresql \
-  lib \
-  bin/initdb \
-  bin/pg_ctl \
-  bin/postgres
+  rm -rf $PKG_DIR/linux && mkdir -p $PKG_DIR/linux
+  tar -xzf $LINUX_DIST -C $PKG_DIR/linux
+  cd $PKG_DIR/linux/pgsql
+  tar -cJvf $TRG_DIR/postgres-linux-x86_64.txz \
+    share/postgresql \
+    lib \
+    bin/initdb \
+    bin/pg_ctl \
+    bin/postgres
+
+  rm -rf $PKG_DIR/linux_x32 && mkdir -p $PKG_DIR/linux_x32
+  tar -xzf $LINUX_X32_DIST -C $PKG_DIR/linux_x32
+  cd $PKG_DIR/linux_x32/pgsql
+  tar -cJvf $TRG_DIR/postgres-linux-x86_32.txz \
+    share/postgresql \
+    lib \
+    bin/initdb \
+    bin/pg_ctl \
+    bin/postgres
+
+fi
